CompressedBeliefMDPs.jl

Documentation for CompressedBeliefMDPs.jl

CompressedBeliefMDPs.CompressedBeliefMDP — Type

struct CompressedBeliefMDP{B, A} <: MDP{B, A}

A struct representing a Markov Decision Process (MDP) with compressed beliefs.

Fields

bmdp::GenerativeBeliefMDP: The underlying GenerativeBeliefMDP that defines the original MDP.

compressor::Compressor: The compressor used to compress beliefs in the MDP.

Type Parameters

B: Type of the belief space.

A: Type of the action space.

Example

pomdp = TigerPOMDP()
updater = DiscreteUpdater(pomdp)
compressor = PCACompressor(1)
generative_mdp = GenerativeBeliefMDP(pomdp, updater)
mdp = CompressedBeliefMDP(generative_mdp, compressor)

CompressedBeliefMDPs.CompressedBeliefMDP — Method

CompressedBeliefMDP(pomdp::POMDP, updater::Updater, compressor::Compressor)

Create a CompressedBeliefMDP based on a given POMDP, updater, and compressor.

This function initializes a GenerativeBeliefMDP using the provided pomdp and updater. It then constructs a CompressedBeliefMDP with the specified compressor.

Arguments

pomdp::POMDP: The original partially observable Markov decision process (POMDP).

updater::Updater: The belief updater used in the GenerativeBeliefMDP.

compressor::Compressor: The compressor used to compress beliefs in the MDP.

Returns

CompressedBeliefMDP{CompressedBeliefMDPState, actiontype(bmdp)}: The compressed belief MDP.

Example

pomdp = MyPOMDP()  # replace with your specific POMDP type
updater = MyUpdater()  # replace with your specific updater type
compressor = MyCompressor()  # replace with your specific compressor type

mdp = CompressedBeliefMDP(pomdp, updater, compressor)

CompressedBeliefMDPs.CompressedBeliefMDPState — Type

struct CompressedBeliefMDPState

A struct representing the state of a CompressedBeliefMDP.

Fields

b̃::AbstractArray{<:Real}: Compressed belief vector.

CompressedBeliefMDPs.Compressor — Type

Base type for an MDP/POMDP belief compression.

CompressedBeliefMDPs.compress — Function

compress(compressor::Compressor, beliefs)

Compress the sampled beliefs using method associated with compressor, and returns a compressed representation.

CompressedBeliefMDPs.decompress — Function

decompress(compressor::Compressor, compressed)

Decompress the compressed beliefs using method associated with compressor, and returns the reconstructed beliefs.

CompressedBeliefMDPs.fit! — Function

fit!(compressor::Compressor, beliefs)

Fit the compressor to beliefs.

CompressedBeliefMDPs.sample — Function

sample(sampler::Sampler, pomdp::POMDP; n_samples::Integer=100)

Return a matrix of beliefs sampled from pomdp.