Implemented Compressors

CompressedBeliefMDPs currently provides wrappers for the following compression types:

a principal component analysis (PCA) compressor,

a kernel PCA compressor,

a probabilistic PCA compressor,

a factor analysis compressor,

an isomap compressor,

an autoencoder compressor

Principal Component Analysis (PCA)

CompressedBeliefMDPs.PCACompressor — Function

Wrapper for MultivariateStats.PCA.

Kernel PCA

CompressedBeliefMDPs.KernelPCACompressor — Function

Wrapper for MultivariateStats.KernelPCA.

Probabilistic PCA

CompressedBeliefMDPs.PPCACompressor — Function

Wrapper for MultivariateStats.PPCA.

Factor Analysis

CompressedBeliefMDPs.FactorAnalysisCompressor — Function

Wrapper for MultivariateStats.FactorAnalysis

Isomap

CompressedBeliefMDPs.IsomapCompressor — Function

Wrapper for ManifoldLearning.Isomap.

Warning

Some compression algorithms aren't optimized for large belief spaces. While they pass our unit tests, they may fail on large POMDPs or without seeding. For large POMDPs, users may want a custom Compressor.